buffer (kø) til nyeste punker, for and undgå self-collision.

+++ src/main/java/org/team08/pspacessnake/Server/GameLogic.java
@@ -63,7 +63,7 @@ public class GameLogic {
 	}
 
 	public boolean playerIsOnBoard(Point point) {
-		return (point.getX() < 0d && point.getY() > (double) gameSettings.getWidth() && point.getY() < 0d && point.getY() > gameSettings.getHeight());
+		return (0d <= point.getX() && point.getX() <= gameSettings.getWidth() && 0d <= point.getY() && point.getY() <= gameSettings.getHeight());
 	}
 
 /*
@@ -115,6 +115,17 @@ public class GameLogic {
 		return false;
 	}
 	
+	public boolean checkBufferedPointsCollision(Player currentPlayer) {
+		for (Player thisPlayer : players) {
+			if (thisPlayer == currentPlayer) continue;
+			for (Point bufferPoint : thisPlayer.getPointBuffer())
+				if (currentPlayer.getPosition().distance(bufferPoint) < currentPlayer.getPosition().getRadius() + bufferPoint.getRadius()) {
+					return true;
+				}
+			}
+		return false;
+	}
+	
 	public int getCellX(Point point) {
 		return (int) point.getX() / gameSettings.getCellSize();
 	}
@@ -124,10 +135,17 @@ public class GameLogic {
 	}
 
 	/*
-	 * Adds new player position point to appropriate cell.
+	 * Adds new player position to buffer and releases non-overlapping buffered points to appropriate cell.  
 	 */
-	public void addPoint(Point point) {
-		boardCellsIterators[getCellY(point)][getCellX(point)].add(point);
+	public void addPoint(Player currentPlayer) {
+		Point playerPos = currentPlayer.getPosition();
+		for (Point bufferPoint : currentPlayer.getPointBuffer()) {
+			if (playerPos.distance(bufferPoint) > playerPos.getRadius() + bufferPoint.getRadius()) {
+				boardCellsIterators[getCellY(bufferPoint)][getCellX(bufferPoint)].add(bufferPoint);
+				currentPlayer.getPointBuffer().remove(bufferPoint);
+			}
+		}
+		currentPlayer.getPointBuffer().add(currentPlayer.getPosition());
 	}
 
 	public GameSettings getGameSettings() {
@@ -139,7 +157,6 @@ public class GameLogic {
 			this.players = new ArrayList<>();
 		}
 		this.players.add(player);
-		i++;
 	}
 
 	public boolean isStarted() {
@@ -162,19 +179,23 @@ public class GameLogic {
 			}
 		}
 	}
+	
+	
 
 	public List<Player> nextFrame() {
 		for (Player player : players) {
 			//if (player.isDead()) continue;	// We might just remove dead player from players ???
 			player.turn();
-			// player.move(gameSettings.getWidth(), gameSettings.getHeight());
-			//Point newPoint = player.move(gameSettings.getWidth(), gameSettings.getHeight());
-			/*if (!playerIsOnBoard(newPoint)) {
+			//player.move(gameSettings.getWidth(), gameSettings.getHeight());
+			//player.move();
+			// Point newPoint = player.move(gameSettings.getWidth(), gameSettings.getHeight());
+			Point newPoint = player.move();
+			if (!playerIsOnBoard(newPoint)) {
 				player.kill();
 				continue;
 			}
-			addPoint(newPoint);*/
-			addPoint(player.move());
+			addPoint(player);
+			//addPoint(player.getPosition());
 			// if (checkCollision(newPoint)) player.kill();
 			
 		}
@@ -187,29 +208,16 @@ public class GameLogic {
 				player.setRemember(holes);	
 			}
 		}
-
 	}
 
 	public Player makePlayer(Token token) {
         Player newPlayer = new Player(token);
         Point newPoint = new Point(ThreadLocalRandom.current().nextInt(5, gameSettings.getWidth() - 5), ThreadLocalRandom.current().nextInt(5, gameSettings.getHeight() - 5), colorList[i]);
-        // Point newPoint = new Point(ThreadLocalRandom.current().nextInt(0, 80), ThreadLocalRandom.current().nextInt(0, 100), colorList[i]);
         newPlayer.setColor(colorList[i]);
         newPlayer.setPosition(newPoint);
         newPlayer.setAngle(newPlayer.getPosition().getAngleToPoint(gameSettings.getWidth() / 2d, gameSettings.getHeight() / 2d));
-        System.out.printf("START: [%f, %f]\tAngle: %f rad\t MIDT: [%f, %f] ", newPoint.getX(), newPoint.getY(), newPlayer.getAngle(), gameSettings.getWidth() / 2d, gameSettings.getHeight() / 2d);
+        //System.out.printf("START: [%f, %f]\tAngle: %f rad\t MIDT: [%f, %f] ", newPoint.getX(), newPoint.getY(), newPlayer.getAngle(), gameSettings.getWidth() / 2d, gameSettings.getHeight() / 2d);
         // i++;
         return newPlayer;
     }
-
-/*
-	public Player makePlayer(Token token) {
-		Player newPlayer = new Player(token);
-		Point newPoint = new Point(ThreadLocalRandom.current().nextInt(0, 80), ThreadLocalRandom.current().nextInt(0, 100), colorList[i]);
-		newPlayer.setColor(colorList[i]);
-		newPlayer.setPosition(newPoint);
-		i++;
-		return newPlayer;
-
-	}*/
 }
